pyproject.toml
[tool.poetry]
name = "surepay-api"
version = "0.1.0"
description = "Surepay FastAPI backend (starter)"
authors = ["Surepay <eng@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = {version = "^0.30.0", extras = ["standard"]}
pydantic = "^2.7.0"
python-jose = "^3.3.0"
structlog = "^24.1.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
redis = "^5.0.0"
httpx = "^0.27.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.6"
xlsxwriter = "^3.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
pytest = "^8.3.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.12.0"
responses = "^0.24.0"
factory-boy = "^3.3.0"
freezegun = "^1.4.0"
faker = "^20.0.0"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests that take more than 10 seconds",
    "security: Security-related tests",
    "payment: Payment processing tests",
    "terminal: Terminal management tests",
    "webhook: Webhook system tests",
    "monitoring: Monitoring and analytics tests",
    "fraud: Fraud detection tests",
    "rbac: Role-based access control tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

